Fix for module resolution and typescript disagreeing
When importing escalade in an ESM module, typescript incorrectly determines it's type and attributes the functionality to the "default" import.

However, it appears that escalade properly supports ESM modules and so the bare import has the call signature. This is a workaround to satisfy the type checker and getting correct inference.